requirements.txt
# local package
-e .

# external requirements
numpy
keras
tensorflow
pandas
click
coverage
flake8
python-dotenv>=0.5.1
